% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_module.R
\name{dateUI}
\alias{dateUI}
\alias{dateServer}
\title{date module}
\usage{
dateUI(id)

dateServer(id)
}
\arguments{
\item{id}{A string that defines the namespace of this module. Make sure this
is the same when calling the UI and the Server function.}
}
\value{
Returns a reactive vector with the selected date inputs.
}
\description{
This module allows the user to select the dates to filter table results.
}
\section{Functions}{
\itemize{
\item \code{dateServer()}: Server function.

}}
