% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_elrsurveiller.R
\name{elrsurveiller}
\alias{elrsurveiller}
\title{Main function to launch the app}
\usage{
elrsurveiller(...)
}
\arguments{
\item{...}{Pass through parameters for shiny.}
}
\description{
\code{elrsurveiller} is the main function that launches the app.
}
