% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Main function to launch the app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Pass through parameters for shiny.}
}
\description{
\code{elrsurveiller} is the main function that launches the app.
}
